monitoring/instrument/requirements.txt
# FastAPI server
fastapi==0.96.0
uvicorn[standard]==0.22.0
joblib==1.1.0
scikit-learn==1.0.2
pandas==1.4.4
lightgbm==3.3.4

# Logging
loguru==0.7.0

# Prometheus metrics
prometheus-client==0.17.1
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-prometheus==1.12.0rc1

# Jaeger tracing
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0

# MinIO client
minio==7.2.7
